Add files via upload
The original Thornthwaite method (Thornthwaite 1948) was developed to estimate monthly potential evapotranspiration (PET) for a standard month with 30 days and an average 12 hour daily photoperiod (daylength) based on monthly mean of daily average air temperatures (Tavg).
Willmott et al. (1985) adjusted the Thornthwaite PET to account for months with days other than 30 and for days with other than 12 hour daylength. Also, an alternative formula was proposed for the cases when Tavg > 26.5 C in the given month as this was known to be a limit of applicability of the original formula.
Camargo et al. (1999) suggested to use effective temperature (Teff) instead of Tavg.
Pereira et al. (2004) used the so far modified method to estimate daily PET by modifying Teff to account for the PET differences between days that have the same Teff but different photoperiods. 

This Python prg will calculate the Thermal index of the year: